Do not delete current user in development

+++ src/app/core/services/auth.service.ts
@@ -6,6 +6,7 @@ import { BsModalService } from 'ngx-bootstrap/modal/';
 import { ConfirmModalComponent } from '../confirm-modal/confirm-modal.component';
 import { Router } from '@angular/router';
 import { StorageMap } from '@ngx-pwa/local-storage';
+import { environment } from 'src/environments/environment';
 
 
 @Injectable({
@@ -31,9 +32,9 @@ export class AuthService {
   };
 
   constructor(private adalService: AdalService,
-              private modalService: BsModalService,
-              private storage: StorageMap,
-              private _router: Router) { this.adalService.init(this.adalConfig); }
+    private modalService: BsModalService,
+    private storage: StorageMap,
+    private _router: Router) { this.adalService.init(this.adalConfig); }
 
   public isLoggedIn(): boolean {
     return this.adalService.userInfo.authenticated;
@@ -42,7 +43,9 @@ export class AuthService {
   public signout(): void {
     this.confirmSignOut().subscribe(res => {
       if (res) {
-        this.storage.delete('currentUser').subscribe();
+        if (environment.production) {
+          this.storage.delete('currentUser').subscribe();
+        }
         localStorage.setItem('prev', this._router.url);
         this.adalService.logOut();
       }
